LayoutTests/svg/custom/use-nested-missing-target-removed-expected.txt

CONSOLE MESSAGE: line 7: TypeError: Result of expression 'document.getElementById("myText1")' [null] is not an object.
PASS
